move global states to user-swr
